fix: 🐛 rename to singular folder name
