+++ design-toolkit/components/src/components/query-builder/__fixture__/example-configuration.ts
